Finish smooth week/day integration

+++ Android/AndroidStudio/RHSMustangs/app/src/main/java/com/sarangjoshi/rhsmustangs/content/SPeriod.java
@@ -1,6 +1,5 @@
 package com.sarangjoshi.rhsmustangs.content;
 
-import com.parse.ParseException;
 import com.parse.ParseObject;
 
 import java.util.Calendar;
@@ -163,10 +162,10 @@ public class SPeriod {
         public STime(int nH, int nM) {
             hour = nH % 24;
             if (minute >= 60) {
-                hour += (int) (minute / 60);
+                hour += minute / 60;
                 minute = nM % 60;
             } else if (minute < 0) {
-                hour += (int) (minute / 60);
+                hour += minute / 60;
                 minute = nM % 60;
             } else
                 minute = nM;
